pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes-webui"
version = "1.0.0"
description = "Process monitoring and configuration UI for Hermes Agent"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "pyyaml",
    "python-dotenv",
    "websockets",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]

[tool.setuptools.packages.find]
include = ["webui", "webui.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
